+++ packages/xrpl/src/models/ledger/FeeSettings.ts
@@ -1,4 +1,4 @@
-import { BaseLedgerEntry, MissingPreviousTxnID } from './BaseLedgerEntry'
+import { BaseLedgerEntry, HasOptionalPreviousTxnID } from './BaseLedgerEntry'
 
 /**
  * The unique id for the FeeSettings object https://xrpl.org/feesettings.html#feesettings-id-format
@@ -26,7 +26,9 @@ export interface FeeSettingsPostAmendmentFields {
   ReserveIncrementDrops: string
 }
 
-export interface FeeSettingsBase extends BaseLedgerEntry, MissingPreviousTxnID {
+export interface FeeSettingsBase
+  extends BaseLedgerEntry,
+    HasOptionalPreviousTxnID {
   LedgerEntryType: 'FeeSettings'
   /**
    * A bit-map of boolean flags for this object. No flags are defined for this type.