+++ packages/xrpl/HISTORY.md
@@ -10,6 +10,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ### Added
 * Add `nfts_by_issuer` clio-only API definition
+* Support for the `fixPreviousTxnID` amendment.
 
 ## 3.1.0 (2024-06-03)